requirements.txt
requests~=2.31.0
jmespath~=1.0.1
PyExecJS~=1.5.1
ddddocr~=1.4.11
pillow~=10.2.0